🔨 simplify code, standardize cf error handling
