+++ baker/siteRenderers.tsx
@@ -7,9 +7,9 @@ import { StaticCollectionPage } from "../site/collections/StaticCollectionPage.j
 import NotFoundPage from "../site/NotFoundPage.js"
 import { DonatePage } from "../site/DonatePage.js"
 import { ExplorerIndexPage } from "../site/ExplorerIndexPage.js"
-import { SubscribePage } from "../site/SubscribePage.js"
-import { OldSubscribePage } from "../site/OldSubscribePage.js"
-import { EmailNotificationsPreferencesPage } from "../site/EmailNotificationsPreferencesPage.js"
+import { SubscribePage } from "../site/Newsletter/SubscribePage.js"
+import { OldSubscribePage } from "../site/Newsletter/OldSubscribePage.js"
+import { EmailNotificationsPreferencesPage } from "../site/Newsletter/EmailNotificationsPreferencesPage.js"
 import { ThankYouPage } from "../site/ThankYouPage.js"
 import TombstonePage from "../site/TombstonePage.js"
 import OwidGdocPage from "../site/gdocs/OwidGdocPage.js"
