--- site/css/email-notifications-subscribe-form.scss
@@ -1,305 +1,0 @@
-.email-notifications-subscribe-form__success {
-    color: $blue-90;
-}
-
-.email-notifications-subscribe-form__newsletter {
-    display: flex;
-    gap: 24px;
-    align-items: flex-start;
-
-    // The illustrations hang into the empty grid columns to the left of the
-    // form, so the checkboxes align with the fieldsets below.
-    @include lg-up {
-        margin-left: -109px;
-    }
-
-    .checkbox input:checked + .custom {
-        background: $blue-70;
-        border-color: $blue-70;
-    }
-}
-
-.email-notifications-subscribe-form__newsletter-image {
-    width: 85px;
-    height: 46px;
-    flex-shrink: 0;
-
-    @include sm-only {
-        display: none;
-    }
-}
-
-.email-notifications-subscribe-form__newsletter-title {
-    @include body-2-bold-tight;
-    color: $blue-90;
-}
-
-.email-notifications-subscribe-form__newsletter-cadence {
-    @include note-12-medium;
-    color: $blue-60;
-    margin-left: 2px;
-}
-
-.email-notifications-subscribe-form__newsletter-description {
-    @include body-3-medium;
-    color: $blue-90;
-    margin: 4px 0 0;
-}
-
-.email-notifications-subscribe-form__divider {
-    width: 100%;
-    margin: 24px 0;
-    border: none;
-    border-top: 1px solid $blue-20;
-}
-
-.email-notifications-subscribe-form__fieldset {
-    margin: 32px 0 0;
-    padding: 0;
-    border: none;
-
-    legend {
-        margin-bottom: 16px;
-        color: $blue-60;
-    }
-}
-
-.email-notifications-subscribe-form__hint {
-    color: $blue-60;
-    margin: 8px 0 0;
-}
-
-.email-notifications-subscribe-form__pills {
-    display: flex;
-    flex-wrap: wrap;
-    align-items: center;
-    gap: 8px;
-}
-
-.email-notifications-subscribe-form__pill {
-    @include body-3-medium;
-    display: inline-flex;
-    align-items: center;
-    gap: 4px;
-    height: 32px;
-    padding: 4px 12px;
-    border: 1px solid $blue-20;
-    border-radius: 1000px;
-    background: none;
-    color: $blue-60;
-    cursor: pointer;
-
-    svg {
-        font-size: 0.75rem;
-    }
-
-    &:hover {
-        background-color: $gray-5;
-    }
-}
-
-.email-notifications-subscribe-form__pill--selected {
-    background-color: $accent-pale-blue;
-    border-color: $accent-pale-blue;
-    color: $blue-90;
-
-    &:hover {
-        background-color: $accent-pale-blue;
-    }
-}
-
-.email-notifications-subscribe-form__select-all {
-    @include body-3-medium;
-    display: inline-flex;
-    align-items: center;
-    gap: 4px;
-    margin-top: 8px;
-    padding: 4px 0;
-    border: none;
-    background: none;
-    color: $blue-90;
-    cursor: pointer;
-
-    svg {
-        font-size: 0.75rem;
-    }
-
-    span {
-        text-decoration: underline;
-    }
-}
-
-.email-notifications-subscribe-form__cards {
-    display: grid;
-    grid-template-columns: 1fr;
-    gap: 8px;
-
-    @include md-up {
-        grid-template-columns: 1fr 1fr;
-        gap: 16px;
-    }
-}
-
-.email-notifications-subscribe-form__card {
-    display: flex;
-    flex-direction: column;
-    align-items: flex-start;
-    width: 100%;
-    padding: 12px;
-    border: 1px solid $blue-20;
-    background: none;
-    text-align: left;
-    cursor: pointer;
-
-    &:hover {
-        background-color: $gray-5;
-    }
-}
-
-.email-notifications-subscribe-form__card--selected {
-    background-color: $accent-pale-blue;
-    border-color: $accent-pale-blue;
-
-    &:hover {
-        background-color: $accent-pale-blue;
-    }
-}
-
-.email-notifications-subscribe-form__card-header {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    width: 100%;
-    color: $blue-60;
-    font-size: 0.75rem;
-}
-
-.email-notifications-subscribe-form__checkbox {
-    border: 1px solid $blue-20;
-    border-radius: 3px;
-    background: #fff;
-    width: 16px;
-    height: 16px;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
-
-.email-notifications-subscribe-form__checkbox--selected {
-    border-color: $blue-60;
-    background: $blue-60;
-    svg {
-        color: #fff;
-    }
-}
-
-.email-notifications-subscribe-form__card-title {
-    @include body-3-medium;
-    margin-top: 6px;
-    color: $blue-90;
-}
-
-.email-notifications-subscribe-form__card-description {
-    @include body-3-medium;
-    color: $blue-60;
-    margin-top: 2px;
-}
-
-.email-notifications-subscribe-form__frequency-options {
-    display: flex;
-    flex-wrap: wrap;
-    gap: 8px;
-}
-
-.email-notifications-subscribe-form__frequency-option {
-    @include body-3-medium;
-    position: relative;
-    display: flex;
-    align-items: center;
-    height: 40px;
-    padding: 8px 16px;
-    border: 1px solid $blue-20;
-    color: $blue-50;
-    cursor: pointer;
-
-    input {
-        position: absolute;
-        inset: 0;
-        margin: 0;
-        opacity: 0;
-        cursor: pointer;
-    }
-
-    &:hover {
-        background-color: $gray-5;
-    }
-
-    &:focus-within {
-        outline: 2px solid $blue-60;
-        outline-offset: 1px;
-    }
-}
-
-.email-notifications-subscribe-form__frequency-option--selected {
-    background-color: $accent-pale-blue;
-    border-color: $accent-pale-blue;
-    color: $blue-90;
-    font-weight: 600;
-
-    &:hover {
-        background-color: $accent-pale-blue;
-    }
-}
-
-.email-notifications-subscribe-form__alert {
-    margin-top: 16px;
-    color: $vermillion;
-}
-
-.email-notifications-subscribe-form__email-submit {
-    display: flex;
-    gap: 8px;
-    margin-top: 40px;
-    padding-top: 24px;
-    border-top: 1px solid $blue-20;
-}
-
-.email-notifications-subscribe-form__email {
-    @include body-3-medium;
-    flex: 1;
-    padding: 8px 16px;
-    border: 1px solid $blue-20;
-
-    &::placeholder {
-        color: $blue-40;
-    }
-}
-
-.email-notifications-subscribe-form__submit {
-    @include body-3-medium;
-    padding: 8px 24px;
-    background-color: $vermillion;
-    color: $white;
-    border: none;
-    cursor: pointer;
-
-    &:hover {
-        background-color: $accent-vermillion;
-    }
-
-    &:disabled {
-        background-color: $gray-30;
-        cursor: default;
-    }
-}
-
-.email-notifications-subscribe-form__privacy-notice {
-    @include note-12-medium;
-    margin-top: 8px;
-    color: $blue-60;
-
-    a {
-        color: inherit;
-        text-decoration: underline;
-    }
-}
